feat: Implement and integrate an affiliate rebate system, applying rewards after successful top-ups.

+++ controller/task.go
@@ -168,7 +168,7 @@ func updateSunoTaskAll(ctx context.Context, channelId int, taskIds []string, tas
 					if err != nil {
 						logger.LogError(ctx, "fail to increase user quota: "+err.Error())
 					}
-					logContent := fmt.Sprintf("异步任务执行失败 %s，补偿 %s", task.TaskID, logger.LogQuota(quota))
+					logContent := fmt.Sprintf("异步任务执行失败 %s，补偿 %s", task.TaskID, logger.LogQuota(int64(quota)))
 					model.RecordLog(task.UserId, model.LogTypeSystem, logContent)
 				}
 			}
